Only support scalar expression constants on PHP 5.6 and up
This checks the PHP version before adding more allowed tokens to the
ReflectionConstant class, to avoid using tokens not available in older
versions of PHP. Since the support for scalar expressions for constants
was added in PHP 5.6 anyway, this does no harm.

Also adds PHP 5.5 and 5.6 to the test matrix in .travis.yml and enables
the container based infrastructure there.